chore(README.md): update table in action inputs section to include new 'context' input feat(action.yml): add 'context' input to action.yml file to allow more context to be added to the generated description feat(autofill_description.py): add support for '--context' command line argument to provide additional context for the generated description
The changes were made to enhance the functionality of the GitHub Action. The 'context' input was added to the action inputs in the README.md file to allow users to provide more context for the generated description. The 'context' input was also added to the action.yml file to define the input in the action metadata. In the autofill_description.py script, support for the '--context' command line argument was added to capture the additional context provided by the user. This additional context can be used to better describe the changes that took place in the generated pull request description.